1.lekcia/FirstReactAgent/pyproject.toml
[project]
name = "firstreactagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "load-dotenv>=0.1.0",
    "openai>=1.102.0",
    "requests>=2.32.5",
]
